move orderby component
