+++ src/ServicePulse.Host/vue/src/components/failedmessages/FailedMessageGroups.vue
@@ -8,7 +8,7 @@ import LicenseExpired from "../../components/LicenseExpired.vue";
 import ServiceControlNotAvailable from "../ServiceControlNotAvailable.vue";
 import LastTenOperations from "../failedmessages/LastTenOperations.vue";
 import MessageGroupList, { IMessageGroupList } from "../failedmessages/MessageGroupList.vue";
-import OrderBy, { getSortFunction } from "./OrderBy.vue";
+import OrderBy, { getSortFunction } from "@/components/OrderBy.vue";
 import SortOptions, { SortDirection } from "@/resources/SortOptions";
 import GroupOperation from "@/resources/GroupOperation";
