+++ src/ServicePulse.Host/vue/src/views/HeartbeatsView.vue
@@ -7,7 +7,7 @@ import routeLinks from "@/router/routeLinks";
 import isRouteSelected from "@/composables/isRouteSelected";
 import { DisplayType, sortOptions, useHeartbeatsStore } from "@/stores/HeartbeatsStore";
 import { storeToRefs } from "pinia";
-import OrderBy from "@/components/failedmessages/OrderBy.vue";
+import OrderBy from "@/components/OrderBy.vue";
 
 const store = useHeartbeatsStore();
 const { inactiveEndpoints, activeEndpoints, selectedDisplay, filterString } = storeToRefs(store);
